Treat conda R more like homebrew R
