+++ extensions/positron-r/src/provider.ts
@@ -17,7 +17,7 @@ import { LOGGER } from './extension';
 import { EXTENSION_ROOT_DIR, MINIMUM_R_VERSION } from './constants';
 import { getInterpreterOverridePaths, printInterpreterSettingsInfo, userRBinaries, userRHeadquarters } from './interpreter-settings.js';
 import { isDirectory, isFile } from './path-utils.js';
-import { isCondaAvailable, getCondaEnvironments, getCondaRPaths, getCondaName } from './provider-conda.js';
+import { discoverCondaBinaries } from './provider-conda.js';
 
 // We don't give this a type so it's compatible with both the VS Code
 // and the LSP types
@@ -45,6 +45,7 @@ export enum RRuntimeSource {
 	system = 'System',
 	user = 'User',
 	homebrew = 'Homebrew',
+	conda = 'Conda',
 }
 
 /**
@@ -233,21 +234,14 @@ export async function makeMetadata(
 	// it's a Homebrew installation if it does)
 	const isHomebrewInstallation = rInst.binpath.includes('/homebrew/');
 
-	const runtimeSource = isHomebrewInstallation ? RRuntimeSource.homebrew :
-		isUserInstallation ?
-			RRuntimeSource.user : RRuntimeSource.system;
+	const isCondaInstallation = rInst.reasonDiscovered && rInst.reasonDiscovered.includes(ReasonDiscovered.CONDA);
 
-	let runtimeShortName: string;
-	runtimeShortName = includeArch ? `${rInst.version} (${rInst.arch})` : rInst.version;
-	// Short name shown to users (when disambiguating within a language)
-	if (rInst.reasonDiscovered && rInst.reasonDiscovered.includes(ReasonDiscovered.CONDA)) {
-		const condaName = getCondaName(rInst.homepath);
-		if (condaName === "") {
-			runtimeShortName = `${runtimeShortName} (Conda)`; // in case no conda name is detected
-		} else {
-			runtimeShortName = `${runtimeShortName} (Conda: ${condaName})`;
-		}
-	}
+	const runtimeSource =
+		isHomebrewInstallation ? RRuntimeSource.homebrew :
+			isCondaInstallation ? RRuntimeSource.conda :
+				isUserInstallation ? RRuntimeSource.user : RRuntimeSource.system;
+
+	const runtimeShortName = includeArch ? `${rInst.version} (${rInst.arch})` : rInst.version;
 
 	// Full name shown to users
 	const runtimeName = `R ${runtimeShortName}`;
@@ -599,39 +593,6 @@ function discoverAdHocBinaries(paths: string[]): RBinary[] {
  */
 function discoverSystemBinaries(): RBinary[] {
 	return discoverHQBinaries(rHeadquarters());
-}
-
-/**
- * Discovers R binaries that are installed in conda environments.
- * @returns conda R binaries.
- */
-async function discoverCondaBinaries(): Promise<RBinary[]> {
-	if (!(await isCondaAvailable())) {
-		LOGGER.info('Conda is not installed or not in PATH.');
-		return [];
-	}
-
-	const condaEnvs = await getCondaEnvironments();
-	const rBinaries: RBinary[] = [];
-
-	for (const envPath of condaEnvs) {
-		const rPaths = getCondaRPaths(envPath);  // list of R binaries in this environment
-
-		if (rPaths.length === 0) {
-			continue;
-		}
-
-		for (const rPath of rPaths) {
-			if (fs.existsSync(rPath)) { // return the first existing R
-				LOGGER.info(`Detected R in Conda environment: ${rPath}`);
-				rBinaries.push({ path: rPath, reasons: [ReasonDiscovered.CONDA] });
-				break;
-			}
-		}
-
-	}
-
-	return rBinaries;
 }
 
 /**
